+++ Assets/Scripts/Battle/Bot/API/BotAPI.cs
@@ -1,55 +1,58 @@
-using CoreSumo;
+using SumoCore;
 using UnityEngine;
 
-public class BotAPI
+namespace SumoBot
 {
-    // [Todo]: Should change the SumoController to exact attribute, giving SumoController directly might be risky
-    // Need to configure asmdef (assembly scope)
-    internal readonly SumoController myController;
-    internal readonly Transform enemyTransform;
-
-    public BotAPI(SumoController myController, Transform enemyTransform)
+    public class BotAPI
     {
-        this.myController = myController;
-        this.enemyTransform = enemyTransform;
-    }
+        // [Todo]: Should change the SumoController to exact attribute, giving SumoController directly might be risky
+        // Need to configure asmdef (assembly scope)
+        internal readonly SumoController myController;
+        internal readonly Transform enemyTransform;
 
-    public Transform MyTransform
-    {
-        get
+        public BotAPI(SumoController myController, Transform enemyTransform)
         {
-            return myController.transform;
+            this.myController = myController;
+            this.enemyTransform = enemyTransform;
         }
-    }
 
-    public Transform EnemyTransform
-    {
-        get
+        public Transform MyTransform
         {
-            return enemyTransform;
+            get
+            {
+                return myController.transform;
+            }
         }
-    }
 
-    public Vector2 GetLastVelocity()
-    {
-        return myController.LastVelocity;
-    }
-
-    public SumoSkill GetSkillAttr()
-    {
-        return myController.Skill;
-    }
-
-    public bool CanExecute(ISumoAction action)
-    {
-        return Controller.InputProvider.CanExecute(action);
-    }
-
-    public SumoController Controller
-    {
-        get
+        public Transform EnemyTransform
         {
-            return myController;
+            get
+            {
+                return enemyTransform;
+            }
+        }
+
+        public Vector2 GetLastVelocity()
+        {
+            return myController.LastVelocity;
+        }
+
+        public SumoSkill GetSkillAttr()
+        {
+            return myController.Skill;
+        }
+
+        public bool CanExecute(ISumoAction action)
+        {
+            return Controller.InputProvider.CanExecute(action);
+        }
+
+        public SumoController Controller
+        {
+            get
+            {
+                return myController;
+            }
         }
     }
 }
