+++ Assets/Scripts/Battle/Bot/Example.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61741bd067c184c968d0d42754d63502
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
