+++ Assets/Scripts/Battle/Bot/Example/EA.meta
