+++ Assets/Scripts/Battle/Bot/Example/EA/EA_Basic_Data.cs.meta
