+++ Assets/Scripts/Battle/Bot/Example/Other.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 608cc6e4d228940e28aae2f210e0c0ee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
