+++ Assets/Scripts/Battle/Input/ButtonPointerHandler.cs
@@ -1,41 +1,43 @@
-using System;
+using SumoHelper;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
-
-public class ButtonPointerHandler : MonoBehaviour, IPointerDownHandler, IPointerUpHandler
+namespace SumoInput
 {
-    #region Action properties
-    public ActionRegistry Actions = new();
-    static public string ActionOnHold = "ActionOnHold";
-    static public string ActionOnPress = "ActionOnPress";
-    private bool isHolding = false;
-    #endregion
-
-    #region Unity methods
-    public void OnPointerDown(PointerEventData eventData)
+    public class ButtonPointerHandler : MonoBehaviour, IPointerDownHandler, IPointerUpHandler
     {
-        isHolding = true;
-    }
+        #region Action properties
+        public ActionRegistry Actions = new();
+        static public string ActionOnHold = "ActionOnHold";
+        static public string ActionOnPress = "ActionOnPress";
+        private bool isHolding = false;
+        #endregion
 
-    public void OnPointerUp(PointerEventData eventData)
-    {
-        isHolding = false;
-        if (GetComponent<Button>().interactable)
+        #region Unity methods
+        public void OnPointerDown(PointerEventData eventData)
         {
-            Actions[ActionOnPress]?.Invoke();
+            isHolding = true;
         }
-    }
 
-    private void Update()
-    {
-        if (isHolding)
+        public void OnPointerUp(PointerEventData eventData)
         {
+            isHolding = false;
             if (GetComponent<Button>().interactable)
             {
-                Actions[ActionOnHold]?.Invoke();
+                Actions[ActionOnPress]?.Invoke();
             }
         }
+
+        private void Update()
+        {
+            if (isHolding)
+            {
+                if (GetComponent<Button>().interactable)
+                {
+                    Actions[ActionOnHold]?.Invoke();
+                }
+            }
+        }
+        #endregion
     }
-    #endregion
 }
