+++ Assets/Scripts/Battle/Log/CollisionLog.cs
@@ -1,21 +1,22 @@
 
 using System.Collections.Generic;
 
-
-public class CollisionLog : BaseLog
+namespace SumoLog
 {
-    public bool IsActor;
-    public float Impact;
-    public float BounceResistance;
-    public float LockDuration = float.NaN;
-    public bool IsSkillActive;
-    public bool IsDashActive;
-
-    public float Duration;
-
-    public new Dictionary<string, dynamic> ToMap()
+    public class CollisionLog : BaseLog
     {
-        return new()
+        public bool IsActor;
+        public float Impact;
+        public float BounceResistance;
+        public float LockDuration = float.NaN;
+        public bool IsSkillActive;
+        public bool IsDashActive;
+
+        public float Duration;
+
+        public new Dictionary<string, dynamic> ToMap()
+        {
+            return new()
         {
             { "IsActor", IsActor},
             { "Impact", Impact},
@@ -26,5 +27,6 @@ public class CollisionLog : BaseLog
             { "Duration", Duration},
             { "Robot", base.ToMap()},
         };
+        }
     }
 }
