+++ Assets/Scripts/Battle/Managers/PauseMenuManager.cs
@@ -1,38 +1,41 @@
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-public class PausePanelManager : MonoBehaviour
+namespace SumoManager
 {
-    public GameObject PausePanel;
-
-    public void ShowPause()
+    public class PausePanelManager : MonoBehaviour
     {
-        PausePanel.SetActive(true);
-        Time.timeScale = 0;
-    }
+        public GameObject PausePanel;
 
-    public void OnResume()
-    {
-        PausePanel.SetActive(false);
-        Time.timeScale = 1;
-    }
+        public void ShowPause()
+        {
+            PausePanel.SetActive(true);
+            Time.timeScale = 0;
+        }
 
-    public void OnRestart()
-    {
-        Time.timeScale = 1;
-        SceneManager.LoadScene(SceneManager.GetActiveScene().name);
-    }
+        public void OnResume()
+        {
+            PausePanel.SetActive(false);
+            Time.timeScale = 1;
+        }
 
-    public void OnBack()
-    {
-        Time.timeScale = 1;
-        SceneManager.LoadScene("campaignScene");
-    }
+        public void OnRestart()
+        {
+            Time.timeScale = 1;
+            SceneManager.LoadScene(SceneManager.GetActiveScene().name);
+        }
+
+        public void OnBack()
+        {
+            Time.timeScale = 1;
+            SceneManager.LoadScene("campaignScene");
+        }
 
 
-    public void OnOut()
-    {
-        Time.timeScale = 1;
-        SceneManager.LoadScene("MainMenu");
+        public void OnOut()
+        {
+            Time.timeScale = 1;
+            SceneManager.LoadScene("MainMenu");
+        }
     }
 }
