+++ Assets/Scripts/Helper/EventLogger.cs
@@ -1,163 +1,168 @@
 using System;
-using CoreSumo;
+using SumoCore;
+using SumoLog;
+using SumoManager;
 using UnityEngine;
 
-public class EventLogger
+namespace SumoHelper
 {
-    public bool IsActive;
-    public bool ForceSave = true;
-
-    private float debounceTime;
-    private float lastCallTime;
-
-    private float startTime;
-    private CollisionLog collision;
-    private readonly ActionLog action;
-    private readonly SumoController controller;
-
-    public EventLogger(SumoController controller, float debounceTime, bool forceSave = true)
+    public class EventLogger
     {
-        this.controller = controller;
-        this.debounceTime = debounceTime;
-        ForceSave = forceSave;
-        action = new();
-    }
+        public bool IsActive;
+        public bool ForceSave = true;
 
-    public EventLogger(SumoController controller, bool forceSave)
-    {
-        ForceSave = forceSave;
-        this.controller = controller;
-    }
+        private float debounceTime;
+        private float lastCallTime;
 
-    public void Call(ISumoAction action)
-    {
-        if (!IsActive)
+        private float startTime;
+        private CollisionLog collision;
+        private readonly ActionLog action;
+        private readonly SumoController controller;
+
+        public EventLogger(SumoController controller, float debounceTime, bool forceSave = true)
         {
-            IsActive = true;
-            startTime = BattleManager.Instance.ElapsedTime;
-
-            this.action.Action = action;
-            this.action.Position = controller.transform.position;
-            this.action.Rotation = controller.transform.rotation.eulerAngles.z;
-            this.action.LinearVelocity = controller.LastVelocity;
-            this.action.AngularVelocity = controller.LastAngularVelocity;
-
-            SaveAction(true);
+            this.controller = controller;
+            this.debounceTime = debounceTime;
+            ForceSave = forceSave;
+            action = new();
         }
 
-        lastCallTime = BattleManager.Instance.ElapsedTime;
-    }
-
-    public void Call(CollisionLog collision = null)
-    {
-        if (this.collision == null && collision != null)
+        public EventLogger(SumoController controller, bool forceSave)
         {
-            this.collision = collision;
-            debounceTime = collision.LockDuration;
+            ForceSave = forceSave;
+            this.controller = controller;
         }
 
-        if (!IsActive)
+        public void Call(ISumoAction action)
         {
-            IsActive = true;
-            startTime = BattleManager.Instance.ElapsedTime;
+            if (!IsActive)
+            {
+                IsActive = true;
+                startTime = BattleManager.Instance.ElapsedTime;
 
-            collision.Position = controller.transform.position;
-            collision.Rotation = controller.transform.rotation.eulerAngles.z;
-            collision.LinearVelocity = controller.LastVelocity;
-            collision.AngularVelocity = controller.LastAngularVelocity;
+                this.action.Action = action;
+                this.action.Position = controller.transform.position;
+                this.action.Rotation = controller.transform.rotation.eulerAngles.z;
+                this.action.LinearVelocity = controller.LastVelocity;
+                this.action.AngularVelocity = controller.LastAngularVelocity;
 
-            SaveCollision(true);
+                SaveAction(true);
+            }
+
+            lastCallTime = BattleManager.Instance.ElapsedTime;
         }
 
-        lastCallTime = BattleManager.Instance.ElapsedTime;
-    }
+        public void Call(CollisionLog collision = null)
+        {
+            if (this.collision == null && collision != null)
+            {
+                this.collision = collision;
+                debounceTime = collision.LockDuration;
+            }
 
-    public void Update()
-    {
-        if (IsActive && BattleManager.Instance.ElapsedTime - lastCallTime >= debounceTime)
+            if (!IsActive)
+            {
+                IsActive = true;
+                startTime = BattleManager.Instance.ElapsedTime;
+
+                collision.Position = controller.transform.position;
+                collision.Rotation = controller.transform.rotation.eulerAngles.z;
+                collision.LinearVelocity = controller.LastVelocity;
+                collision.AngularVelocity = controller.LastAngularVelocity;
+
+                SaveCollision(true);
+            }
+
+            lastCallTime = BattleManager.Instance.ElapsedTime;
+        }
+
+        public void Update()
+        {
+            if (IsActive && BattleManager.Instance.ElapsedTime - lastCallTime >= debounceTime)
+            {
+                IsActive = false;
+
+                if (action != null)
+                    SaveAction(false);
+                else if (collision != null)
+                    SaveCollision(false);
+            }
+        }
+
+        public void ForceStopAndSave()
         {
             IsActive = false;
-
             if (action != null)
                 SaveAction(false);
-            else if (collision != null)
+            if (collision != null)
                 SaveCollision(false);
         }
-    }
 
-    public void ForceStopAndSave()
-    {
-        IsActive = false;
-        if (action != null)
-            SaveAction(false);
-        if (collision != null)
-            SaveCollision(false);
-    }
-
-    public void SaveAction(bool isStart = false)
-    {
-        if (!isStart)
+        public void SaveAction(bool isStart = false)
         {
-            action.Rotation = controller.transform.rotation.eulerAngles.z;
-            action.Position = controller.transform.position;
-            action.LinearVelocity = controller.LastVelocity;
-            action.AngularVelocity = controller.LastAngularVelocity;
-
-            float duration = BattleManager.Instance.ElapsedTime - startTime;
-            action.Duration = duration;
-        }
-
-        LogManager.LogPlayerEvents(
-            actor: controller.Side,
-            startedAt: startTime,
-            isStart: isStart,
-            category: "Action",
-            data: action.ToMap()
-        );
-    }
-
-    public void SaveCollision(bool isStart = false)
-    {
-        Debug.Log("SaveCollision");
-
-        PlayerSide? target = null;
-
-        if (!isStart)
-        {
-            collision.Rotation = controller.transform.rotation.eulerAngles.z;
-            collision.Position = controller.transform.position;
-            collision.LinearVelocity = controller.LastVelocity;
-            collision.AngularVelocity = controller.LastAngularVelocity;
-
-            float duration = BattleManager.Instance.ElapsedTime - startTime;
-            collision.Duration = duration;
-        }
-        else
-        {
-            if (collision.IsActor)
+            if (!isStart)
             {
-                if (controller.Side == PlayerSide.Left)
-                {
-                    target = PlayerSide.Right;
-                }
-                else
-                {
-                    target = PlayerSide.Left;
-                }
+                action.Rotation = controller.transform.rotation.eulerAngles.z;
+                action.Position = controller.transform.position;
+                action.LinearVelocity = controller.LastVelocity;
+                action.AngularVelocity = controller.LastAngularVelocity;
+
+                float duration = BattleManager.Instance.ElapsedTime - startTime;
+                action.Duration = duration;
+            }
+
+            LogManager.LogPlayerEvents(
+                actor: controller.Side,
+                startedAt: startTime,
+                isStart: isStart,
+                category: "Action",
+                data: action.ToMap()
+            );
+        }
+
+        public void SaveCollision(bool isStart = false)
+        {
+            Debug.Log("SaveCollision");
+
+            PlayerSide? target = null;
+
+            if (!isStart)
+            {
+                collision.Rotation = controller.transform.rotation.eulerAngles.z;
+                collision.Position = controller.transform.position;
+                collision.LinearVelocity = controller.LastVelocity;
+                collision.AngularVelocity = controller.LastAngularVelocity;
+
+                float duration = BattleManager.Instance.ElapsedTime - startTime;
+                collision.Duration = duration;
             }
             else
             {
-                target = null;
+                if (collision.IsActor)
+                {
+                    if (controller.Side == PlayerSide.Left)
+                    {
+                        target = PlayerSide.Right;
+                    }
+                    else
+                    {
+                        target = PlayerSide.Left;
+                    }
+                }
+                else
+                {
+                    target = null;
+                }
             }
-        }
 
-        LogManager.LogPlayerEvents(
-            actor: controller.Side,
-            target: target,
-            startedAt: startTime,
-            isStart: isStart,
-            category: "Collision",
-            data: collision.ToMap()
-        );
+            LogManager.LogPlayerEvents(
+                actor: controller.Side,
+                target: target,
+                startedAt: startTime,
+                isStart: isStart,
+                category: "Collision",
+                data: collision.ToMap()
+            );
+        }
     }
 }
