+++ Assets/Scripts/Helper/TrackableAction.cs
@@ -1,56 +1,59 @@
 using System;
 using System.Collections.Generic;
 
-[Serializable]
-public class TrackableAction
+namespace SumoHelper
 {
-    [NonSerialized]
-    private Action<object[]> action;
-
-    [NonSerialized]
-    private readonly HashSet<Delegate> subscribers = new();
-
-    public void Subscribe(Action<object[]> callback)
+    [Serializable]
+    public class TrackableAction
     {
-        if (subscribers.Add(callback))
+        [NonSerialized]
+        private Action<object[]> action;
+
+        [NonSerialized]
+        private readonly HashSet<Delegate> subscribers = new();
+
+        public void Subscribe(Action<object[]> callback)
         {
-            action += callback;
+            if (subscribers.Add(callback))
+            {
+                action += callback;
+            }
         }
-    }
 
-    public void Unsubscribe(Action<object[]> callback)
-    {
-        if (subscribers.Remove(callback))
+        public void Unsubscribe(Action<object[]> callback)
         {
-            action -= callback;
+            if (subscribers.Remove(callback))
+            {
+                action -= callback;
+            }
         }
-    }
 
-    public void Invoke(object[] param)
-    {
-        action?.Invoke(param);
-    }
-
-    public void Invoke(object arg)
-    {
-        Invoke(new object[] { arg });
-    }
-
-    public void Invoke()
-    {
-        Invoke(new object[] { });
-    }
-
-    public IReadOnlyCollection<Delegate> Subscribers => subscribers;
-
-    public int SubscribersCount => subscribers.Count;
-    public IEnumerable<string> GetSubscriberDescriptions()
-    {
-        foreach (var d in subscribers)
+        public void Invoke(object[] param)
         {
-            var method = d.Method;
-            var target = d.Target;
-            yield return $"{target?.GetType().Name ?? "Static"}.{method.Name}";
+            action?.Invoke(param);
+        }
+
+        public void Invoke(object arg)
+        {
+            Invoke(new object[] { arg });
+        }
+
+        public void Invoke()
+        {
+            Invoke(new object[] { });
+        }
+
+        public IReadOnlyCollection<Delegate> Subscribers => subscribers;
+
+        public int SubscribersCount => subscribers.Count;
+        public IEnumerable<string> GetSubscriberDescriptions()
+        {
+            foreach (var d in subscribers)
+            {
+                var method = d.Method;
+                var target = d.Target;
+                yield return $"{target?.GetType().Name ?? "Static"}.{method.Name}";
+            }
         }
     }
 }
